requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cleanup==6.0.0
mccabe==0.7.0
Pillow==9.3.0
python-dotenv==0.21.0
pytz==2022.4
sorl-thumbnail==12.9.0
sqlparse==0.4.3
django-ckeditor==6.5.1
django-debug-toolbar==3.7.0